step3 - refactor rename package
