+++ src/main/java/study/racing/RacingInput.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.racing;
 
 public class RacingInput {
