+++ src/main/java/study/racing/ResultView.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.racing;
 
 import java.util.List;
 import java.util.stream.IntStream;
