+++ src/test/java/study/racing/CarRacingTest.java
@@ -1,4 +1,4 @@
-package study.step3;
+package study.racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
